Use separate mapping for scripting due to different type names
