+++ src/openrct2/libopenrct2.vcxproj
@@ -560,45 +560,45 @@
     <ClInclude Include="scenes\title\TitleSequence.h" />
     <ClInclude Include="scenes\title\TitleSequenceManager.h" />
     <ClInclude Include="scenes\title\TitleSequencePlayer.h" />
+    <ClInclude Include="scripting\bindings\entity\ScEntity.hpp" />
     <ClInclude Include="scripting\bindings\entity\ScGuest.hpp" />
     <ClInclude Include="scripting\bindings\entity\ScLitter.hpp" />
     <ClInclude Include="scripting\bindings\entity\ScParticle.hpp" />
     <ClInclude Include="scripting\bindings\entity\ScPeep.hpp" />
     <ClInclude Include="scripting\bindings\entity\ScStaff.hpp" />
     <ClInclude Include="scripting\bindings\entity\ScVehicle.hpp" />
-    <ClInclude Include="scripting\bindings\game\ScPlugin.hpp" />
-    <ClInclude Include="scripting\bindings\game\ScProfiler.hpp" />
-    <ClInclude Include="scripting\bindings\network\ScPlayer.hpp" />
-    <ClInclude Include="scripting\bindings\network\ScPlayerGroup.hpp" />
-    <ClInclude Include="scripting\bindings\object\ScInstalledObject.hpp" />
-    <ClInclude Include="scripting\bindings\object\ScObjectManager.h" />
-    <ClInclude Include="scripting\bindings\ride\ScRideStation.hpp" />
-    <ClInclude Include="scripting\bindings\ride\ScTrackIterator.h" />
-    <ClInclude Include="scripting\bindings\ride\ScTrackSegment.h" />
-    <ClInclude Include="scripting\bindings\world\ScParkMessage.hpp" />
-    <ClInclude Include="scripting\bindings\world\ScResearch.hpp" />
-    <ClInclude Include="scripting\bindings\world\ScTileElement.hpp" />
-    <ClInclude Include="scripting\Duktape.hpp" />
-    <ClInclude Include="scripting\IconNames.hpp" />
-    <ClInclude Include="scripting\HookEngine.h" />
-    <ClInclude Include="scripting\Plugin.h" />
     <ClInclude Include="scripting\bindings\game\ScCheats.hpp" />
-    <ClInclude Include="scripting\bindings\world\ScClimate.hpp" />
     <ClInclude Include="scripting\bindings\game\ScConfiguration.hpp" />
     <ClInclude Include="scripting\bindings\game\ScConsole.hpp" />
     <ClInclude Include="scripting\bindings\game\ScContext.hpp" />
-    <ClInclude Include="scripting\bindings\world\ScDate.hpp" />
     <ClInclude Include="scripting\bindings\game\ScDisposable.hpp" />
-    <ClInclude Include="scripting\bindings\entity\ScEntity.hpp" />
-    <ClInclude Include="scripting\bindings\world\ScMap.hpp" />
+    <ClInclude Include="scripting\bindings\game\ScPlugin.hpp" />
+    <ClInclude Include="scripting\bindings\game\ScProfiler.hpp" />
     <ClInclude Include="scripting\bindings\network\ScNetwork.hpp" />
-    <ClInclude Include="scripting\bindings\object\ScObject.hpp" />
-    <ClInclude Include="scripting\bindings\world\ScPark.hpp" />
-    <ClInclude Include="scripting\bindings\ride\ScRide.hpp" />
-    <ClInclude Include="scripting\ScriptEngine.h" />
-    <ClInclude Include="scripting\bindings\world\ScScenario.hpp" />
+    <ClInclude Include="scripting\bindings\network\ScPlayer.hpp" />
+    <ClInclude Include="scripting\bindings\network\ScPlayerGroup.hpp" />
     <ClInclude Include="scripting\bindings\network\ScSocket.hpp" />
+    <ClInclude Include="scripting\bindings\object\ScInstalledObject.hpp" />
+    <ClInclude Include="scripting\bindings\object\ScObject.hpp" />
+    <ClInclude Include="scripting\bindings\object\ScObjectManager.h" />
+    <ClInclude Include="scripting\bindings\ride\ScRide.hpp" />
+    <ClInclude Include="scripting\bindings\ride\ScRideStation.hpp" />
+    <ClInclude Include="scripting\bindings\ride\ScTrackIterator.h" />
+    <ClInclude Include="scripting\bindings\ride\ScTrackSegment.h" />
+    <ClInclude Include="scripting\bindings\world\ScClimate.hpp" />
+    <ClInclude Include="scripting\bindings\world\ScDate.hpp" />
+    <ClInclude Include="scripting\bindings\world\ScMap.hpp" />
+    <ClInclude Include="scripting\bindings\world\ScPark.hpp" />
+    <ClInclude Include="scripting\bindings\world\ScParkMessage.hpp" />
+    <ClInclude Include="scripting\bindings\world\ScResearch.hpp" />
+    <ClInclude Include="scripting\bindings\world\ScScenario.hpp" />
     <ClInclude Include="scripting\bindings\world\ScTile.hpp" />
+    <ClInclude Include="scripting\bindings\world\ScTileElement.hpp" />
+    <ClInclude Include="scripting\Duktape.hpp" />
+    <ClInclude Include="scripting\HookEngine.h" />
+    <ClInclude Include="scripting\IconNames.hpp" />
+    <ClInclude Include="scripting\Plugin.h" />
+    <ClInclude Include="scripting\ScriptEngine.h" />
     <ClInclude Include="SpriteIds.h" />
     <ClInclude Include="System.hpp" />
     <ClInclude Include="TrackImporter.h" />
@@ -1095,6 +1095,7 @@
     <ClCompile Include="scripting\bindings\network\ScNetwork.cpp" />
     <ClCompile Include="scripting\bindings\network\ScPlayer.cpp" />
     <ClCompile Include="scripting\bindings\network\ScPlayerGroup.cpp" />
+    <ClCompile Include="scripting\bindings\object\ScInstalledObject.cpp" />
     <ClCompile Include="scripting\bindings\object\ScObjectManager.cpp" />
     <ClCompile Include="scripting\bindings\ride\ScRide.cpp" />
     <ClCompile Include="scripting\bindings\ride\ScRideStation.cpp" />
