+++ src/openrct2/scripting/bindings/object/ScObject.hpp
@@ -23,6 +23,7 @@
 
     #include <memory>
     #include <optional>
+    #include <string_view>
 
 namespace OpenRCT2::Scripting
 {
